some minor fixes, after tests

+++ DataFormats/IRPCDigi/interface/IRPCDigi.h
@@ -33,10 +33,10 @@ private:
   uint16_t strip_;
   int32_t bxLR_;    //BX from low radius FEB
   int32_t bxHR_;    //BX from high radius FEB 
-  int32_t sbxLR_;   //sub-BX from low radius FEB     
-  int32_t sbxHR_;   //sub-BX from high radius FEB  
-  int32_t fineLR_;     //high resolution time rom low radius FEB       
-  int32_t fineHR_;     //high resolution time rom high radius FEB 
+  int8_t sbxLR_;   //sub-BX from low radius FEB     
+  int8_t sbxHR_;   //sub-BX from high radius FEB  
+  int8_t fineLR_;     //high resolution time rom low radius FEB       
+  int8_t fineHR_;     //high resolution time rom high radius FEB 
 };
 
 std::ostream& operator<<(std::ostream& o, const IRPCDigi& digi);
